big fixes and add the streaming API
