+++ Containers/CMakeLists.txt
@@ -1,11 +1,11 @@
 
-add_library(MF_Containers STATIC EXCLUDE_FROM_ALL)
+add_library(MF_Containers STATIC EXCLUDE_FROM_ALL include/MF/Streams.hpp)
 target_include_directories(MF_Containers PUBLIC include)
 target_link_libraries(MF_Containers PUBLIC MF_Commons)
 target_sources(
         MF_Containers
         PRIVATE
-            src/hhh.cpp
+            src/dummy_containers.cpp
         PUBLIC
             include/MF/Array.hpp
 )
